refactor(app): pages gets new import paths
