+++ app/src/pages/Desktop/Devices/DeviceDetails/index.tsx
@@ -3,7 +3,7 @@ import { Navigate, useParams } from 'react-router-dom'
 
 import { ApiHostProvider } from '@opentrons/react-api-client'
 
-import { useSyncRobotClock } from '/app/organisms/Devices/hooks'
+import { useSyncRobotClock } from '/app/organisms/Desktop/Devices/hooks'
 import { getScanning, OPENTRONS_USB } from '/app/redux/discovery'
 import { appShellRequestor } from '/app/redux/shell/remote'
 import { DeviceDetailsComponent } from './DeviceDetailsComponent'
