+++ app/src/pages/Desktop/Devices/DevicesLanding/__tests__/DevicesLanding.test.tsx
@@ -3,8 +3,8 @@ import { vi, it, describe, expect, beforeEach, afterEach } from 'vitest'
 
 import { renderWithProviders } from '/app/__testing-utils__'
 import { i18n } from '/app/i18n'
-import { DevicesEmptyState } from '/app/organisms/Devices/DevicesEmptyState'
-import { RobotCard } from '/app/organisms/Devices/RobotCard'
+import { DevicesEmptyState } from '/app/organisms/Desktop/Devices/DevicesEmptyState'
+import { RobotCard } from '/app/organisms/Desktop/Devices/RobotCard'
 import {
   getScanning,
   getConnectableRobots,
@@ -18,8 +18,8 @@ import {
 } from '/app/redux/discovery/__fixtures__'
 import { DevicesLanding } from '..'
 
-vi.mock('/app/organisms/Devices/DevicesEmptyState')
-vi.mock('/app/organisms/Devices/RobotCard')
+vi.mock('/app/organisms/Desktop/Devices/DevicesEmptyState')
+vi.mock('/app/organisms/Desktop/Devices/RobotCard')
 vi.mock('/app/redux/discovery')
 
 const render = () => {
