+++ app/src/pages/Desktop/Devices/DevicesLanding/index.tsx
@@ -27,8 +27,8 @@ import {
   OPENTRONS_USB,
 } from '/app/redux/discovery'
 import { appShellRequestor } from '/app/redux/shell/remote'
-import { RobotCard } from '/app/organisms/Devices/RobotCard'
-import { DevicesEmptyState } from '/app/organisms/Devices/DevicesEmptyState'
+import { RobotCard } from '/app/organisms/Desktop/Devices/RobotCard'
+import { DevicesEmptyState } from '/app/organisms/Desktop/Devices/DevicesEmptyState'
 import { CollapsibleSection } from '/app/molecules/CollapsibleSection'
 
 import { Divider } from '/app/atoms/structure'
