+++ app/src/pages/Desktop/Devices/RobotSettings/__tests__/RobotSettings.test.tsx
@@ -5,8 +5,8 @@ import { Route, MemoryRouter, Routes } from 'react-router-dom'
 import { renderWithProviders } from '/app/__testing-utils__'
 import { i18n } from '/app/i18n'
 import { RobotSettingsCalibration } from '/app/organisms/RobotSettingsCalibration'
-import { RobotSettingsNetworking } from '/app/organisms/Devices/RobotSettings/RobotSettingsNetworking'
-import { RobotSettingsAdvanced } from '/app/organisms/Devices/RobotSettings/RobotSettingsAdvanced'
+import { RobotSettingsNetworking } from '/app/organisms/Desktop/Devices/RobotSettings/RobotSettingsNetworking'
+import { RobotSettingsAdvanced } from '/app/organisms/Desktop/Devices/RobotSettings/RobotSettingsAdvanced'
 import { useRobot } from '/app/redux-resources/robots'
 import { RobotSettings } from '..'
 import { when } from 'vitest-when'
@@ -18,8 +18,8 @@ import {
 import { getRobotUpdateSession } from '/app/redux/robot-update'
 
 vi.mock('/app/organisms/RobotSettingsCalibration')
-vi.mock('/app/organisms/Devices/RobotSettings/RobotSettingsNetworking')
-vi.mock('/app/organisms/Devices/RobotSettings/RobotSettingsAdvanced')
+vi.mock('/app/organisms/Desktop/Devices/RobotSettings/RobotSettingsNetworking')
+vi.mock('/app/organisms/Desktop/Devices/RobotSettings/RobotSettingsAdvanced')
 vi.mock('/app/redux-resources/robots')
 vi.mock('/app/redux/discovery/selectors')
 vi.mock('/app/redux/robot-update')
